requirements.txt
openai>=1.0
anthropic>=0.25
python-dotenv
